feat: refactor user service methods for consistency and update controller methods for improved clarity
